+++ petroleum/src/common/macros.rs
@@ -937,7 +937,7 @@ macro_rules! page_flags_const {
 /// Integrated identity mapping macro
 #[macro_export]
 macro_rules! map_identity_range_macro {
-    ($mapper:expr, $frame_allocator:expr, $start_addr:expr, $pages:expr, $flags:expr) => {{ unsafe { map_identity_range($mapper, $frame_allocator, $start_addr, $pages, $flags) } }};
+    ($mapper:expr, $frame_allocator:expr, $start_addr:expr, $pages:expr, $flags:expr) => {{ unsafe { $crate::page_table::utils::map_identity_range($mapper, $frame_allocator, $start_addr, $pages, $flags) } }};
 }
 
 //// Identity mapping with detailed logging
